@@ -3,16 +3,16 @@ import 'package:flutter/widgets.dart';
 import 'modular_navigator.dart';
 
 class ModularNavigatorOutlet extends ModularNavigator {
-  final NavigatorState global;
-  final NavigatorState module;
+  final NavigatorState? global;
+  final NavigatorState? module;
   ModularNavigatorOutlet({this.module, this.global}) : super(module);
 
   @override
-  void pop<T extends Object>([T result]) {
-    if (module.canPop()) {
-      module.pop();
+  void pop<T extends Object?>([T? result]) {
+    if (module!.canPop()) {
+      module!.pop();
     } else {
-      global.pop();
+      global!.pop();
     }
   }
 }
