@@ -1,9 +1,10 @@
 import 'package:flutter/widgets.dart';
+
 import '../../flutter_modular.dart';
 
 class RouteLink extends IModularNavigator {
-  final String path;
-  final String modulePath;
+  final String? path;
+  final String? modulePath;
 
   RouteLink({this.path, this.modulePath = "/"});
 
@@ -15,64 +16,48 @@ class RouteLink extends IModularNavigator {
   bool canPop() => Modular.navigator.canPop();
 
   @override
-  Future<bool> maybePop<T extends Object>([T result]) {
+  Future<bool> maybePop<T extends Object?>([T? result]) {
     return Modular.navigator.maybePop(result);
   }
 
   @override
-  void pop<T extends Object>([T result]) => Modular.navigator.pop(result);
+  void pop<T extends Object?>([T? result]) => Modular.navigator.pop(result);
 
   @override
-  Future<T> popAndPushNamed<T extends Object, TO extends Object>(
-          String routeName,
-          {TO result,
-          Object arguments}) =>
-      Modular.navigator.popAndPushNamed(_checkpath(routeName),
-          result: result, arguments: arguments);
+  Future<T?> popAndPushNamed<T extends Object?, TO extends Object?>(String routeName,
+          {TO? result, Object? arguments}) =>
+      Modular.navigator.popAndPushNamed(_checkpath(routeName), result: result, arguments: arguments);
 
   @override
-  void popUntil(bool Function(Route) predicate) =>
-      Modular.navigator.popUntil(predicate);
+  void popUntil(bool Function(Route) predicate) => Modular.navigator.popUntil(predicate);
 
   @override
-  Future<T> push<T extends Object>(Route<T> route) =>
-      Modular.navigator.push(route);
+  Future<T?> push<T extends Object?>(Route<T> route) => Modular.navigator.push(route);
 
   @override
-  Future<T> pushNamed<T extends Object>(String routeName, {Object arguments}) =>
+  Future<T?> pushNamed<T extends Object?>(String routeName, {Object? arguments}) =>
       Modular.navigator.pushNamed(_checkpath(routeName), arguments: arguments);
 
   @override
-  Future<T> pushNamedAndRemoveUntil<T extends Object>(
-          String newRouteName, bool Function(Route) predicate,
-          {Object arguments}) =>
-      Modular.navigator.pushNamedAndRemoveUntil(
-          _checkpath(newRouteName), predicate,
-          arguments: arguments);
+  Future<T?> pushNamedAndRemoveUntil<T extends Object?>(String newRouteName, bool Function(Route) predicate,
+          {Object? arguments}) =>
+      Modular.navigator.pushNamedAndRemoveUntil(_checkpath(newRouteName), predicate, arguments: arguments);
   @override
-  Future<T> pushReplacementNamed<T extends Object, TO extends Object>(
-          String routeName,
-          {TO result,
-          Object arguments}) =>
-      Modular.navigator.pushReplacementNamed(_checkpath(routeName),
-          result: result, arguments: arguments);
+  Future<T?> pushReplacementNamed<T extends Object?, TO extends Object?>(String routeName,
+          {TO? result, Object? arguments}) =>
+      Modular.navigator.pushReplacementNamed(_checkpath(routeName), result: result, arguments: arguments);
 
   @override
-  Future<T> pushReplacement<T extends Object, TO extends Object>(
-          Route<T> newRoute,
-          {TO result}) =>
+  Future<T?> pushReplacement<T extends Object?, TO extends Object?>(Route<T> newRoute, {TO? result}) =>
       Modular.navigator.pushReplacement(newRoute, result: result);
 
   @override
   Future showDialog({
-    @deprecated Widget child,
-    @required WidgetBuilder builder,
+    @deprecated Widget? child,
+    WidgetBuilder? builder,
     bool barrierDismissible = true,
   }) =>
-      Modular.navigator.showDialog(
-          builder: builder,
-          child: child,
-          barrierDismissible: barrierDismissible);
+      Modular.navigator.showDialog(builder: builder, child: child, barrierDismissible: barrierDismissible);
 
   String _checkpath(String routeName) {
     routeName = routeName[0] == '/' ? routeName : '/$routeName';
@@ -81,5 +66,5 @@ class RouteLink extends IModularNavigator {
   }
 
   @override
-  NavigatorState get navigator => Modular.to.navigator;
+  NavigatorState? get navigator => Modular.to.navigator;
 }
