@@ -1,8 +1,8 @@
 import '../../flutter_modular.dart';
 
 class Old {
-  final ModularArguments args;
-  final RouteLink link;
+  final ModularArguments? args;
+  final RouteLink? link;
 
   Old({this.args, this.link});
 }
