@@ -3,5 +3,5 @@ import '../../flutter_modular.dart';
 
 abstract class ModularStatelessWidget<T extends ChildModule>
     extends StatelessWidget with InjectWidgetMixin<T> {
-  ModularStatelessWidget({Key key}) : super(key: key);
+  ModularStatelessWidget({Key? key}) : super(key: key);
 }
