@@ -1,14 +1,16 @@
 import 'package:flutter_modular/flutter_modular.dart';
 import 'package:flutter_test/flutter_test.dart';
+import 'package:mockito/annotations.dart';
 import 'package:mockito/mockito.dart';
 
-class NavigatorMock extends Mock implements RouteLink {}
+import 'route_link_test.mocks.dart';
 
+@GenerateMocks([RouteLink])
 void main() {
-  RouteLink link;
+  late RouteLink link;
 
   setUpAll(() {
-    link = NavigatorMock();
+    link = MockRouteLink();
   });
 
   group('Navigation', () {
