@@ -1,6 +1,6 @@
 //ignore:one_member_abstracts
 abstract class GuardExecutor {
-  void onGuarded(String path, {bool isActive});
+  void onGuarded(String path, {bool? isActive});
 }
 
 abstract class RouteGuard {
