@@ -1,13 +1,13 @@
 import 'package:flutter/widgets.dart';
 
 abstract class IModularNavigator {
-  String get path;
-  String get modulePath;
-  NavigatorState get navigator;
+  String? get path;
+  String? get modulePath;
+  NavigatorState? get navigator;
 
   Future showDialog({
-    Widget child,
-    WidgetBuilder builder,
+    Widget? child,
+    WidgetBuilder? builder,
     bool barrierDismissible = true,
   });
 
@@ -16,7 +16,7 @@ abstract class IModularNavigator {
   /// ```
   /// Modular.to.push(MaterialPageRoute(builder: (context) => HomePage()),);
   /// ```
-  Future<T> push<T extends Object>(Route<T> route);
+  Future<T?> push<T extends Object?>(Route<T> route);
 
   /// Pop the current route off the navigator and navigate to a route.
   ///
@@ -27,10 +27,7 @@ abstract class IModularNavigator {
   /// ```
   /// Modular.to.popAndPushNamed('/home', arguments: 10);
   /// ```
-  Future<T> popAndPushNamed<T extends Object, TO extends Object>(
-      String routeName,
-      {TO result,
-      Object arguments});
+  Future<T?> popAndPushNamed<T extends Object?, TO extends Object?>(String routeName, {TO? result, Object? arguments});
 
   /// Navigate to a route.
   ///
@@ -41,7 +38,7 @@ abstract class IModularNavigator {
   /// ```
   /// Modular.to.pushNamed('/home', arguments: 10);
   /// ```
-  Future<T> pushNamed<T extends Object>(String routeName, {Object arguments});
+  Future<T?> pushNamed<T extends Object?>(String routeName, {Object? arguments});
 
   /// Push the route with the given name onto the navigator that most tightly
   /// encloses the given context, and then remove all the previous routes until
@@ -54,9 +51,8 @@ abstract class IModularNavigator {
   /// ```
   /// Modular.to.pushNamedAndRemoveUntil('/home', ModalRoute.withName('/'), arguments: 10);
   /// ```
-  Future<T> pushNamedAndRemoveUntil<T extends Object>(
-      String newRouteName, bool Function(Route<dynamic>) predicate,
-      {Object arguments});
+  Future<T?> pushNamedAndRemoveUntil<T extends Object?>(String newRouteName, bool Function(Route<dynamic>) predicate,
+      {Object? arguments});
 
   ///Replace the current route of the navigator that most tightly encloses the
   ///given context by pushing the route named routeName and then disposing the
@@ -69,10 +65,8 @@ abstract class IModularNavigator {
   /// ```
   /// Modular.to.pushReplacementNamed('/home', arguments: 10);
   /// ```
-  Future<T> pushReplacementNamed<T extends Object, TO extends Object>(
-      String routeName,
-      {TO result,
-      Object arguments});
+  Future<T?> pushReplacementNamed<T extends Object?, TO extends Object?>(String routeName,
+      {TO? result, Object? arguments});
 
   ///Replace the current route of the navigator that most tightly encloses
   ///the given context by pushing the given route and then disposing
@@ -83,16 +77,14 @@ abstract class IModularNavigator {
   ///   MaterialPageRoute(builder: (context) => HomePage())
   /// );
   /// ```
-  Future<T> pushReplacement<T extends Object, TO extends Object>(
-      Route<T> newRoute,
-      {TO result});
+  Future<T?> pushReplacement<T extends Object?, TO extends Object?>(Route<T> newRoute, {TO? result});
 
   /// Removes the current Route from the stack of routes.
   ///
   /// ```
   /// Modular.to.pop();
   /// ```
-  void pop<T extends Object>([T result]);
+  void pop<T extends Object?>([T? result]);
 
   /// The initial route cannot be popped off the navigator, which implies that
   /// this function returns true only if popping the navigator would not remove
@@ -110,7 +102,7 @@ abstract class IModularNavigator {
   /// ```
   /// Modular.to.maybePop();
   /// ```
-  Future<bool> maybePop<T extends Object>([T result]);
+  Future<bool> maybePop<T extends Object?>([T? result]);
 
   ///Calls pop repeatedly on the navigator that most tightly encloses the given
   ///context until the predicate returns true.
