@@ -1,16 +1,16 @@
+import 'package:collection/collection.dart' show IterableExtension;
 import 'package:flutter/material.dart';
 
 import '../../flutter_modular.dart';
 
-void initModule(ChildModule module,
-    {List<Bind> changeBinds, bool initialModule}) {
+void initModule(ChildModule module, {List<Bind>? changeBinds, bool? initialModule}) {
   Modular.debugMode = false;
-  final list = module.binds;
+  final list = module.binds!;
   final changedList = List<Bind>.from(list);
-  for (var item in list ?? []) {
-    var dep = (changeBinds ?? []).firstWhere((dep) {
+  for (var item in list) {
+    var dep = (changeBinds ?? []).firstWhereOrNull((dep) {
       return item.runtimeType == dep.runtimeType;
-    }, orElse: () => null);
+    });
     if (dep != null) {
       changedList.remove(item);
       changedList.add(dep);
@@ -24,7 +24,7 @@ void initModule(ChildModule module,
   }
 }
 
-void initModules(List<ChildModule> modules, {List<Bind> changeBinds}) {
+void initModules(List<ChildModule> modules, {List<Bind>? changeBinds}) {
   for (var module in modules) {
     initModule(module, changeBinds: changeBinds);
   }
