@@ -6,12 +6,12 @@ import 'page_transition.dart';
 PageRouteBuilder<T> fadeInTransition<T>(
     Widget Function(
   BuildContext,
-  ModularArguments,
+  ModularArguments?,
 )
         builder,
-    ModularArguments args,
+    ModularArguments? args,
     Duration transitionDuration,
-    RouteSettings settings) {
+    RouteSettings? settings) {
   return PageRouteBuilder<T>(
       settings: settings,
       transitionDuration: transitionDuration,
@@ -26,11 +26,8 @@ PageRouteBuilder<T> fadeInTransition<T>(
       });
 }
 
-PageRouteBuilder<T> noTransition<T>(
-    Widget Function(BuildContext, ModularArguments) builder,
-    ModularArguments args,
-    Duration transitionDuration,
-    RouteSettings settings) {
+PageRouteBuilder<T> noTransition<T>(Widget Function(BuildContext, ModularArguments?) builder, ModularArguments? args,
+    Duration transitionDuration, RouteSettings? settings) {
   return PageRouteBuilder(
       settings: settings,
       transitionDuration: transitionDuration,
@@ -39,11 +36,8 @@ PageRouteBuilder<T> noTransition<T>(
       });
 }
 
-PageRouteBuilder<T> rightToLeft<T>(
-    Widget Function(BuildContext, ModularArguments) builder,
-    ModularArguments args,
-    Duration transitionDuration,
-    RouteSettings settings) {
+PageRouteBuilder<T> rightToLeft<T>(Widget Function(BuildContext, ModularArguments?) builder, ModularArguments? args,
+    Duration transitionDuration, RouteSettings? settings) {
   return PageTransition<T>(
     settings: settings,
     duration: transitionDuration,
@@ -54,11 +48,8 @@ PageRouteBuilder<T> rightToLeft<T>(
   );
 }
 
-PageRouteBuilder<T> leftToRight<T>(
-    Widget Function(BuildContext, ModularArguments) builder,
-    ModularArguments args,
-    Duration transitionDuration,
-    RouteSettings settings) {
+PageRouteBuilder<T> leftToRight<T>(Widget Function(BuildContext, ModularArguments?) builder, ModularArguments? args,
+    Duration transitionDuration, RouteSettings? settings) {
   return PageTransition<T>(
     settings: settings,
     duration: transitionDuration,
@@ -69,11 +60,8 @@ PageRouteBuilder<T> leftToRight<T>(
   );
 }
 
-PageRouteBuilder<T> upToDown<T>(
-    Widget Function(BuildContext, ModularArguments) builder,
-    ModularArguments args,
-    Duration transitionDuration,
-    RouteSettings settings) {
+PageRouteBuilder<T> upToDown<T>(Widget Function(BuildContext, ModularArguments?) builder, ModularArguments? args,
+    Duration transitionDuration, RouteSettings? settings) {
   return PageTransition<T>(
     settings: settings,
     duration: transitionDuration,
@@ -84,11 +72,8 @@ PageRouteBuilder<T> upToDown<T>(
   );
 }
 
-PageRouteBuilder<T> downToUp<T>(
-    Widget Function(BuildContext, ModularArguments) builder,
-    ModularArguments args,
-    Duration transitionDuration,
-    RouteSettings settings) {
+PageRouteBuilder<T> downToUp<T>(Widget Function(BuildContext, ModularArguments?) builder, ModularArguments? args,
+    Duration transitionDuration, RouteSettings? settings) {
   return PageTransition<T>(
     settings: settings,
     duration: transitionDuration,
@@ -99,11 +84,8 @@ PageRouteBuilder<T> downToUp<T>(
   );
 }
 
-PageRouteBuilder<T> scale<T>(
-    Widget Function(BuildContext, ModularArguments) builder,
-    ModularArguments args,
-    Duration transitionDuration,
-    RouteSettings settings) {
+PageRouteBuilder<T> scale<T>(Widget Function(BuildContext, ModularArguments?) builder, ModularArguments? args,
+    Duration transitionDuration, RouteSettings? settings) {
   return PageTransition<T>(
     settings: settings,
     duration: transitionDuration,
@@ -114,11 +96,8 @@ PageRouteBuilder<T> scale<T>(
   );
 }
 
-PageRouteBuilder<T> rotate<T>(
-    Widget Function(BuildContext, ModularArguments) builder,
-    ModularArguments args,
-    Duration transitionDuration,
-    RouteSettings settings) {
+PageRouteBuilder<T> rotate<T>(Widget Function(BuildContext, ModularArguments?) builder, ModularArguments? args,
+    Duration transitionDuration, RouteSettings? settings) {
   return PageTransition<T>(
     settings: settings,
     duration: transitionDuration,
@@ -129,11 +108,8 @@ PageRouteBuilder<T> rotate<T>(
   );
 }
 
-PageRouteBuilder<T> size<T>(
-    Widget Function(BuildContext, ModularArguments) builder,
-    ModularArguments args,
-    Duration transitionDuration,
-    RouteSettings settings) {
+PageRouteBuilder<T> size<T>(Widget Function(BuildContext, ModularArguments?) builder, ModularArguments? args,
+    Duration transitionDuration, RouteSettings? settings) {
   return PageTransition<T>(
     settings: settings,
     duration: transitionDuration,
@@ -144,11 +120,8 @@ PageRouteBuilder<T> size<T>(
   );
 }
 
-PageRouteBuilder<T> rightToLeftWithFade<T>(
-    Widget Function(BuildContext, ModularArguments) builder,
-    ModularArguments args,
-    Duration transitionDuration,
-    RouteSettings settings) {
+PageRouteBuilder<T> rightToLeftWithFade<T>(Widget Function(BuildContext, ModularArguments?) builder,
+    ModularArguments? args, Duration transitionDuration, RouteSettings? settings) {
   return PageTransition<T>(
     settings: settings,
     duration: transitionDuration,
@@ -159,11 +132,8 @@ PageRouteBuilder<T> rightToLeftWithFade<T>(
   );
 }
 
-PageRouteBuilder<T> leftToRightWithFade<T>(
-    Widget Function(BuildContext, ModularArguments) builder,
-    ModularArguments args,
-    Duration transitionDuration,
-    RouteSettings settings) {
+PageRouteBuilder<T> leftToRightWithFade<T>(Widget Function(BuildContext, ModularArguments?) builder,
+    ModularArguments? args, Duration transitionDuration, RouteSettings? settings) {
   return PageTransition<T>(
     settings: settings,
     duration: transitionDuration,
