@@ -3,7 +3,7 @@ import '../../flutter_modular.dart';
 
 abstract class ModularState<TWidget extends StatefulWidget, TBind>
     extends State<TWidget> {
-  final TBind controller = Modular.get<TBind>();
+  final TBind? controller = Modular.get<TBind>();
 
   @override
   void dispose() {
@@ -13,7 +13,7 @@ abstract class ModularState<TWidget extends StatefulWidget, TBind>
 }
 
 mixin ModularStateMixin<T extends StatefulWidget, TBind> on State<T> {
-  final TBind controller = Modular.get<TBind>();
+  final TBind? controller = Modular.get<TBind>();
 
   @override
   void dispose() {
