@@ -7,7 +7,7 @@ class ProductWidget extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Container(
-      child: Text(Modular.get<ProductBloc>().testingText),
+      child: Text(Modular.get<ProductBloc>()!.testingText),
     );
   }
 }
