@@ -1,5 +1,5 @@
 abstract class ILocalStorage {
-  Future get(String key);
-  Future put(String key, String value);
-  Future delete(String key);
+  Future? get(String key);
+  Future? put(String key, String value);
+  Future? delete(String key);
 }
