@@ -2,17 +2,17 @@ import 'ilocal_repository.dart';
 
 class LocalStorageSharePreference implements ILocalStorage {
   @override
-  Future delete(String key) {
+  Future? delete(String key) {
     return null;
   }
 
   @override
-  Future get(String key) {
+  Future? get(String key) {
     return null;
   }
 
   @override
-  Future put(String key, String value) {
+  Future? put(String key, String value) {
     return null;
   }
 }
